11 Bowler Street
"Certainly."

"You must confine yourself to your room, on pretence of a headache, when your stepfather comes back.
Then when you hear him retire for the night, you must open the shutters of your window, undo the
hasp, put your lamp there as a signal to us, and then withdraw quietly with everything which you are
likely to want into the room which you used to occupy. I have no doubt that, in spite of the
repairs, you could manage there for one night."

"Oh, yes, easily."

"The rest you will leave in our hands."