79 Glan Road
"'Surely it would be better to take no notice.'

"'No, no, we should have him loitering here always. Kindly turn round and wave him away like that.'

"I did as I was told, and at the same instant Mrs. Rucastle drew down the blind. That was a week
ago, and from that time I have not sat again in the window, nor have I worn the blue dress, nor seen
the man in the road."

"Pray continue," said Holmes. "Your narrative promises to be a most interesting one."

"You will find it rather disconnected, I fear, and there may prove to be little relation between the
different incidents of which I speak. On the very first day that I was at the Copper Beeches, Mr.
Rucastle took me to a small outhouse which stands near the kitchen door. As we approached it I heard
the sharp rattling of a chain, and the sound as of a large animal moving about.

"'Look in here!' said Mr. Rucastle, showing me a slit between two planks. 'Is he not a beauty?'