17 Larch Walk
"I am endeavouring to tell you everything, Mr. Holmes, which may have any bearing upon the case, but
I beg that you will question me upon any point which I do not make clear."

"On the contrary, your statement is singularly lucid."

"I come to a part of my story now in which I should wish to be particularly so. I am not a very
heavy sleeper, and the anxiety in my mind tended, no doubt, to make me even less so than usual.
About two in the morning, then, I was awakened by some sound in the house. It had ceased ere I was
wide awake, but it had left an impression behind it as though a window had gently closed somewhere.
I lay listening with all my ears. Suddenly, to my horror, there was a distinct sound of footsteps
moving softly in the next room. I slipped out of bed, all palpitating with fear, and peeped round
the corner of my dressing-room door.