9 Glendale
"Last Monday I had finished for the day and was dressing in my room above the opium den when I
looked out of my window and saw, to my horror and astonishment, that my wife was standing in the
street, with her eyes fixed full upon me. I gave a cry of surprise, threw up my arms to cover my
face, and, rushing to my confidant, the Lascar, entreated him to prevent anyone from coming up to
me. I heard her voice downstairs, but I knew that she could not ascend. Swiftly I threw off my
clothes, pulled on those of a beggar, and put on my pigments and wig. Even a wife's eyes could not
pierce so complete a disguise. But then it occurred to me that there might be a search in the room,
and that the clothes might betray me. I threw open the window, reopening by my violence a small cut
which I had inflicted upon myself in the bedroom that morning. Then I seized my coat, which was
weighted by the coppers which I had just transferred to it from the leather bag in which I carried
my takings. I hurled it out of the window, and it disappeared into the Thames. The other clothes
would have followed, but at that moment there was a rush of constables up the stair, and a few
minutes after I found, rather, I confess, to my relief, that instead of being identified as Mr.
Neville St. Clair, I was arrested as his murderer.

"I do not know that there is anything else for me to explain. I was determined to preserve my
disguise as long as possible, and hence my preference for a dirty face. Knowing that my wife would
be terribly anxious, I slipped off my ring and confided it to the Lascar at a moment when no
constable was watching me, together with a hurried scrawl, telling her that she had no cause to
fear."

"That note only reached her yesterday," said Holmes.