138 Glan Road
"'I wish to the Lord, Mr. Wilson, that I was a red-headed man.'

"'Why that?' I asks.

"'Why,' says he, 'here's another vacancy on the League of the Red-headed Men. It's worth quite a
little fortune to any man who gets it, and I understand that there are more vacancies than there are
men, so that the trustees are at their wits' end what to do with the money. If my hair would only
change colour, here's a nice little crib all ready for me to step into.'