17 Down Close
"It was too far for me to go, weak and ill as I was. I determined to wait until I got back to town
before telling my story to the police. It was a little past six when I arrived, so I went first to
have my wound dressed, and then the doctor was kind enough to bring me along here. I put the case
into your hands and shall do exactly what you advise."

We both sat in silence for some little time after listening to this extraordinary narrative. Then
Sherlock Holmes pulled down from the shelf one of the ponderous commonplace books in which he placed
his cuttings.