135 Cliff Place
Again a startled look came over the somewhat vacuous face of Miss Mary Sutherland. "Yes, I did bang
out of the house," she said, "for it made me angry to see the easy way in which Mr. Windibank--that
is, my father--took it all. He would not go to the police, and he would not go to you, and so at
last, as he would do nothing and kept on saying that there was no harm done, it made me mad, and I
just on with my things and came right away to you."

"Your father," said Holmes, "your stepfather, surely, since the name is different."

"Yes, my stepfather. I call him father, though it sounds funny, too, for he is only five years and
two months older than myself."

"And your mother is alive?"