18 Larch Walk
"Quite so. Pray proceed with your statement."

"I could not sleep that night. A vague feeling of impending misfortune impressed me. My sister and
I, you will recollect, were twins, and you know how subtle are the links which bind two souls which
are so closely allied. It was a wild night. The wind was howling outside, and the rain was beating
and splashing against the windows. Suddenly, amid all the hubbub of the gale, there burst forth the
wild scream of a terrified woman. I knew that it was my sister's voice. I sprang from my bed,
wrapped a shawl round me, and rushed into the corridor. As I opened my door I seemed to hear a low
whistle, such as my sister described, and a few moments later a clanging sound, as if a mass of
metal had fallen. As I ran down the passage, my sister's door was unlocked, and revolved slowly upon
its hinges. I stared at it horror-stricken, not knowing what was about to issue from it. By the
light of the corridor-lamp I saw my sister appear at the opening, her face blanched with terror, her
hands groping for help, her whole figure swaying to and fro like that of a drunkard. I ran to her
and threw my arms round her, but at that moment her knees seemed to give way and she fell to the
ground.  She writhed as one who is in terrible pain, and her limbs were dreadfully convulsed. At
first I thought that she had not recognised me, but as I bent over her she suddenly shrieked out in
a voice which I shall never forget, 'Oh, my God! Helen! It was the band! The speckled band!' There
was something else which she would fain have said, and she stabbed with her finger into the air in
the direction of the doctor's room, but a fresh convulsion seized her and choked her words. I rushed
out, calling loudly for my stepfather, and I met him hastening from his room in his dressing-gown.
When he reached my sister's side she was unconscious, and though he poured brandy down her throat
and sent for medical aid from the village, all efforts were in vain, for she slowly sank and died
without having recovered her consciousness. Such was the dreadful end of my beloved sister."

"One moment," said Holmes, "are you sure about this whistle and metallic sound? Could you swear to
it?"

"That was what the county coroner asked me at the inquiry. It is my strong impression that I heard
it, and yet, among the crash of the gale and the creaking of an old house, I may possibly have been
deceived."

"Was your sister dressed?"

"No, she was in her night-dress. In her right hand was found the charred stump of a match, and in
her left a match-box."