16 Wellswood Gardens
The question was hardly out of my mouth, and Holmes had not yet opened his lips to reply, when we
heard a heavy footfall in the passage and a tap at the door.

"This is the girl's stepfather, Mr. James Windibank," said Holmes. "He has written to me to say that
he would be here at six. Come in!"

The man who entered was a sturdy, middle-sized fellow, some thirty years of age, clean-shaven, and
sallow-skinned, with a bland, insinuating manner, and a pair of wonderfully sharp and penetrating
grey eyes. He shot a questioning glance at each of us, placed his shiny top-hat upon the sideboard,
and with a slight bow sidled down into the nearest chair.

"Good-evening, Mr. James Windibank," said Holmes. "I think that this typewritten letter is from you,
in which you made an appointment with me for six o'clock?"