176 Sunningdale Drive
"Pray continue," said Holmes. "Your narrative promises to be a most interesting one."

"You will find it rather disconnected, I fear, and there may prove to be little relation between the
different incidents of which I speak. On the very first day that I was at the Copper Beeches, Mr.
Rucastle took me to a small outhouse which stands near the kitchen door. As we approached it I heard
the sharp rattling of a chain, and the sound as of a large animal moving about.

"'Look in here!' said Mr. Rucastle, showing me a slit between two planks. 'Is he not a beauty?'

"I looked through and was conscious of two glowing eyes, and of a vague figure huddled up in the
darkness.

"'Don't be frightened,' said my employer, laughing at the start which I had given. 'It's only Carlo,
my mastiff. I call him mine, but really old Toller, my groom, is the only man who can do anything
with him. We feed him once a day, and not too much then, so that he is always as keen as mustard.
Toller lets him loose every night, and God help the trespasser whom he lays his fangs upon. For
goodness' sake don't you ever on any pretext set your foot over the threshold at night, for it's as
much as your life is worth.'

"The warning was no idle one, for two nights later I happened to look out of my bedroom window about
two o'clock in the morning. It was a beautiful moonlight night, and the lawn in front of the house
was silvered over and almost as bright as day. I was standing, rapt in the peaceful beauty of the
scene, when I was aware that something was moving under the shadow of the copper beeches. As it
emerged into the moonshine I saw what it was. It was a giant dog, as large as a calf, tawny tinted,
with hanging jowl, black muzzle, and huge projecting bones. It walked slowly across the lawn and
vanished into the shadow upon the other side. That dreadful sentinel sent a chill to my heart which
I do not think that any burglar could have done.