25 Glan Road
"I could hardly imagine a more damning case," I remarked. "If ever circumstantial evidence pointed
to a criminal it does so here."

"Circumstantial evidence is a very tricky thing," answered Holmes thoughtfully. "It may seem to
point very straight to one thing, but if you shift your own point of view a little, you may find it
pointing in an equally uncompromising manner to something entirely different. It must be confessed,
however, that the case looks exceedingly grave against the young man, and it is very possible that
he is indeed the culprit. There are several people in the neighbourhood, however, and among them
Miss Turner, the daughter of the neighbouring landowner, who believe in his innocence, and who have
retained Lestrade, whom you may recollect in connection with the Study in Scarlet, to work out the
case in his interest. Lestrade, being rather puzzled, has referred the case to me, and hence it is
that two middle-aged gentlemen are flying westward at fifty miles an hour instead of quietly
digesting their breakfasts at home."