22 Valentia Road
"If I can be of use."

"Oh, a trusty comrade is always of use; and a chronicler still more so. My room at The Cedars is a
double-bedded one."

"The Cedars?"

"Yes; that is Mr. St. Clair's house. I am staying there while I conduct the inquiry."

"Where is it, then?"