115 Larch Walk
"How, in the name of good-fortune, did you know all that, Mr. Holmes?" he asked. "How did you know,
for example, that I did manual labour. It's as true as gospel, for I began as a ship's carpenter."

"Your hands, my dear sir. Your right hand is quite a size larger than your left. You have worked
with it, and the muscles are more developed."

"Well, the snuff, then, and the Freemasonry?"