84 Holmefield Avenue
"Smart fellow, that," observed Holmes as we walked away. "He is, in my judgment, the fourth smartest
man in London, and for daring I am not sure that he has not a claim to be third. I have known
something of him before."

"Evidently," said I, "Mr. Wilson's assistant counts for a good deal in this mystery of the Red-
headed League. I am sure that you inquired your way merely in order that you might see him."

"Not him."

"What then?"