25 Balcombe Close
"Good heavens!" I cried, "this is a terrible injury. It must have bled considerably."

"Yes, it did. I fainted when it was done, and I think that I must have been senseless for a long
time. When I came to I found that it was still bleeding, so I tied one end of my handkerchief very
tightly round the wrist and braced it up with a twig."

"Excellent! You should have been a surgeon."

"It is a question of hydraulics, you see, and came within my own province."

"This has been done," said I, examining the wound, "by a very heavy and sharp instrument."

"A thing like a cleaver," said he.