50 Wellswood Gardens
"Two years have passed since then, and my life has been until lately lonelier than ever. A month
ago, however, a dear friend, whom I have known for many years, has done me the honour to ask my hand
in marriage. His name is Armitage--Percy Armitage--the second son of Mr. Armitage, of Crane Water,
near Reading. My stepfather has offered no opposition to the match, and we are to be married in the
course of the spring. Two days ago some repairs were started in the west wing of the building, and
my bedroom wall has been pierced, so that I have had to move into the chamber in which my sister
died, and to sleep in the very bed in which she slept. Imagine, then, my thrill of terror when last
night, as I lay awake, thinking over her terrible fate, I suddenly heard in the silence of the night
the low whistle which had been the herald of her own death. I sprang up and lit the lamp, but
nothing was to be seen in the room. I was too shaken to go to bed again, however, so I dressed, and
as soon as it was daylight I slipped down, got a dog-cart at the Crown Inn, which is opposite, and
drove to Leatherhead, from whence I have come on this morning with the one object of seeing you and
asking your advice."

"You have done wisely," said my friend. "But have you told me all?"

"Yes, all."

"Miss Roylott, you have not. You are screening your stepfather."