32 Down Close
"But his left-handedness."

"You were yourself struck by the nature of the injury as recorded by the surgeon at the inquest. The
blow was struck from immediately behind, and yet was upon the left side. Now, how can that be unless
it were by a left-handed man? He had stood behind that tree during the interview between the father
and son. He had even smoked there. I found the ash of a cigar, which my special knowledge of tobacco
ashes enables me to pronounce as an Indian cigar. I have, as you know, devoted some attention to
this, and written a little monograph on the ashes of 140 different varieties of pipe, cigar, and
cigarette tobacco. Having found the ash, I then looked round and discovered the stump among the moss
where he had tossed it. It was an Indian cigar, of the variety which are rolled in Rotterdam."

"And the cigar-holder?"