7 Tamworth Drive
"Thank you. I have already learned all that is in the public prints, nothing more. I presume that I
may take it as correct--this article, for example, as to the disappearance of the bride."

Lord St. Simon glanced over it. "Yes, it is correct, as far as it goes."

"But it needs a great deal of supplementing before anyone could offer an opinion. I think that I may
arrive at my facts most directly by questioning you."

"Pray do so."

"When did you first meet Miss Hatty Doran?"

"In San Francisco, a year ago."