6 Larch Walk
"It is fortunate, for you will perhaps be able to post me up. I read nothing except the criminal
news and the agony column. The latter is always instructive. But if you have followed recent events
so closely you must have read about Lord St. Simon and his wedding?"

"Oh, yes, with the deepest interest."

"That is well. The letter which I hold in my hand is from Lord St. Simon. I will read it to you, and
in return you must turn over these papers and let me have whatever bears upon the matter. This is
what he says: