61 Holmefield Avenue
"It is The Morning Chronicle of April 27, 1890. Just two months ago."

"Very good. Now, Mr. Wilson?"

"Well, it is just as I have been telling you, Mr. Sherlock Holmes," said Jabez Wilson, mopping his
forehead; "I have a small pawnbroker's business at Coburg Square, near the City. It's not a very
large affair, and of late years it has not done more than just give me a living. I used to be able
to keep two assistants, but now I only keep one; and I would have a job to pay him but that he is
willing to come for half wages so as to learn the business."