73 Larch Walk
I had called upon my friend Sherlock Holmes upon the second morning after Christmas, with the
intention of wishing him the compliments of the season. He was lounging upon the sofa in a purple
dressing-gown, a pipe-rack within his reach upon the right, and a pile of crumpled morning papers,
evidently newly studied, near at hand. Beside the couch was a wooden chair, and on the angle of the
back hung a very seedy and disreputable hard-felt hat, much the worse for wear, and cracked in
several places. A lens and a forceps lying upon the seat of the chair suggested that the hat had
been suspended in this manner for the purpose of examination.

"You are engaged," said I; "perhaps I interrupt you."

"Not at all. I am glad to have a friend with whom I can discuss my results. The matter is a
perfectly trivial one"--he jerked his thumb in the direction of the old hat--"but there are points
in connection with it which are not entirely devoid of interest and even of instruction."

I seated myself in his armchair and warmed my hands before his crackling fire, for a sharp frost had
set in, and the windows were thick with the ice crystals. "I suppose," I remarked, "that, homely as
it looks, this thing has some deadly story linked on to it--that it is the clue which will guide you
in the solution of some mystery and the punishment of some crime."

"No, no. No crime," said Sherlock Holmes, laughing. "Only one of those whimsical little incidents
which will happen when you have four million human beings all jostling each other within the space
of a few square miles. Amid the action and reaction of so dense a swarm of humanity, every possible
combination of events may be expected to take place, and many a little problem will be presented
which may be striking and bizarre without being criminal. We have already had experience of such."