60 Moor End
"My name is Sherlock Holmes. It is my business to know what other people don't know."

"But you can know nothing of this?"

"Excuse me, I know everything of it. You are endeavouring to trace some geese which were sold by
Mrs. Oakshott, of Brixton Road, to a salesman named Breckinridge, by him in turn to Mr. Windigate,
of the Alpha, and by him to his club, of which Mr. Henry Baker is a member."

"Oh, sir, you are the very man whom I have longed to meet," cried the little fellow with
outstretched hands and quivering fingers. "I can hardly explain to you how interested I am in this
matter."

Sherlock Holmes hailed a four-wheeler which was passing. "In that case we had better discuss it in a
cosy room rather than in this wind-swept market-place," said he. "But pray tell me, before we go
farther, who it is that I have the pleasure of assisting."

The man hesitated for an instant. "My name is John Robinson," he answered with a sidelong glance.