32 Beaconside
"Wait a little, Mr. Holder. We must come back to that. About this girl, Miss Holder. You saw her
return by the kitchen door, I presume?"

"Yes; when I went to see if the door was fastened for the night I met her slipping in. I saw the
man, too, in the gloom."