15 Moor End
"It is possible."

"More than that. It is probable. And now you see the deadly urgency of this new case, and why I
urged young Openshaw to caution. The blow has always fallen at the end of the time which it would
take the senders to travel the distance. But this one comes from London, and therefore we cannot
count upon delay."

"Good God!" I cried. "What can it mean, this relentless persecution?"