93 Holmefield Avenue
"It is true. And yet--Well! I wish she had been of my own station! What a queen she would have
made!" He relapsed into a moody silence, which was not broken until we drew up in Serpentine Avenue.

The door of Briony Lodge was open, and an elderly woman stood upon the steps. She watched us with a
sardonic eye as we stepped from the brougham.