72 Corbet Close
"But how?"

"I am about to be married."

"So I have heard."

"To Clotilde Lothman von Saxe-Meningen, second daughter of the King of Scandinavia. You may know the
strict principles of her family. She is herself the very soul of delicacy. A shadow of a doubt as to
my conduct would bring the matter to an end."

"And Irene Adler?"