14 Valentia Road
"Is it possible?" gasped the banker.

"You then roused his anger by calling him names at a moment when he felt that he had deserved your
warmest thanks. He could not explain the true state of affairs without betraying one who certainly
deserved little enough consideration at his hands. He took the more chivalrous view, however, and
preserved her secret."

"And that was why she shrieked and fainted when she saw the coronet," cried Mr. Holder. "Oh, my God!
what a blind fool I have been! And his asking to be allowed to go out for five minutes! The dear
fellow wanted to see if the missing piece were at the scene of the struggle. How cruelly I have
misjudged him!"

"When I arrived at the house," continued Holmes, "I at once went very carefully round it to observe
if there were any traces in the snow which might help me. I knew that none had fallen since the
evening before, and also that there had been a strong frost to preserve impressions. I passed along
the tradesmen's path, but found it all trampled down and indistinguishable. Just beyond it, however,
at the far side of the kitchen door, a woman had stood and talked with a man, whose round
impressions on one side showed that he had a wooden leg. I could even tell that they had been
disturbed, for the woman had run back swiftly to the door, as was shown by the deep toe and light
heel marks, while Wooden-leg had waited a little, and then had gone away. I thought at the time that
this might be the maid and her sweetheart, of whom you had already spoken to me, and inquiry showed
it was so. I passed round the garden without seeing anything more than random tracks, which I took
to be the police; but when I got into the stable lane a very long and complex story was written in
the snow in front of me.

"There was a double line of tracks of a booted man, and a second double line which I saw with
delight belonged to a man with naked feet. I was at once convinced from what you had told me that
the latter was your son. The first had walked both ways, but the other had run swiftly, and as his
tread was marked in places over the depression of the boot, it was obvious that he had passed after
the other. I followed them up and found they led to the hall window, where Boots had worn all the
snow away while waiting. Then I walked to the other end, which was a hundred yards or more down the
lane. I saw where Boots had faced round, where the snow was cut up as though there had been a
struggle, and, finally, where a few drops of blood had fallen, to show me that I was not mistaken.
Boots had then run down the lane, and another little smudge of blood showed that it was he who had
been hurt. When he came to the highroad at the other end, I found that the pavement had been
cleared, so there was an end to that clue.