80 Holmefield Avenue
"Oh, you must not discourage me, Mr. Holmes. I know that all is well with him. There is so keen a
sympathy between us that I should know if evil came upon him. On the very day that I saw him last he
cut himself in the bedroom, and yet I in the dining-room rushed upstairs instantly with the utmost
certainty that something had happened. Do you think that I would respond to such a trifle and yet be
ignorant of his death?"

"I have seen too much not to know that the impression of a woman may be more valuable than the
conclusion of an analytical reasoner. And in this letter you certainly have a very strong piece of
evidence to corroborate your view. But if your husband is alive and able to write letters, why
should he remain away from you?"

"I cannot imagine. It is unthinkable."

"And on Monday he made no remarks before leaving you?"

"No."

"And you were surprised to see him in Swandam Lane?"