59 Sunningdale Drive
"Your French gold?"

"Yes. We had occasion some months ago to strengthen our resources and borrowed for that purpose
30,000 napoleons from the Bank of France. It has become known that we have never had occasion to
unpack the money, and that it is still lying in our cellar. The crate upon which I sit contains
2,000 napoleons packed between layers of lead foil. Our reserve of bullion is much larger at present
than is usually kept in a single branch office, and the directors have had misgivings upon the
subject."

"Which were very well justified," observed Holmes. "And now it is time that we arranged our little
plans. I expect that within an hour matters will come to a head. In the meantime Mr. Merryweather,
we must put the screen over that dark lantern."