89 Cliff Place
It was close upon four before the door opened, and a drunken-looking groom, ill-kempt and side-
whiskered, with an inflamed face and disreputable clothes, walked into the room. Accustomed as I was
to my friend's amazing powers in the use of disguises, I had to look three times before I was
certain that it was indeed he. With a nod he vanished into the bedroom, whence he emerged in five
minutes tweed-suited and respectable, as of old. Putting his hands into his pockets, he stretched
out his legs in front of the fire and laughed heartily for some minutes.

"Well, really!" he cried, and then he choked and laughed again until he was obliged to lie back,
limp and helpless, in the chair.