153 Glan Road
"Stoke Moran?" said he.

"Yes, sir, that be the house of Dr. Grimesby Roylott," remarked the driver.

"There is some building going on there," said Holmes; "that is where we are going."

"There's the village," said the driver, pointing to a cluster of roofs some distance to the left;
"but if you want to get to the house, you'll find it shorter to get over this stile, and so by the
foot-path over the fields. There it is, where the lady is walking."