87 Tamworth Drive
"It is half-past ten now."

"Yes, and I have no doubt that is her ring."

"It may turn out to be of more interest than you think. You remember that the affair of the blue
carbuncle, which appeared to be a mere whim at first, developed into a serious investigation. It may
be so in this case, also."

"Well, let us hope so. But our doubts will very soon be solved, for here, unless I am much mistaken,
is the person in question."

As he spoke the door opened and a young lady entered the room. She was plainly but neatly dressed,
with a bright, quick face, freckled like a plover's egg, and with the brisk manner of a woman who
has had her own way to make in the world.