34 Balcombe Close
"Yes, sir."

"Or rather, I fancy, of that goose. It was one bird, I imagine in which you were interested--white,
with a black bar across the tail."

Ryder quivered with emotion. "Oh, sir," he cried, "can you tell me where it went to?"

"It came here."

"Here?"

"Yes, and a most remarkable bird it proved. I don't wonder that you should take an interest in it.
It laid an egg after it was dead--the bonniest, brightest little blue egg that ever was seen. I have
it here in my museum."