55 Wilson Gardens
As I entered, a sallow Malay attendant had hurried up with a pipe for me and a supply of the drug,
beckoning me to an empty berth.

"Thank you. I have not come to stay," said I. "There is a friend of mine here, Mr. Isa Whitney, and
I wish to speak with him."

There was a movement and an exclamation from my right, and peering through the gloom, I saw Whitney,
pale, haggard, and unkempt, staring out at me.

"My God! It's Watson," said he. He was in a pitiable state of reaction, with every nerve in a
twitter. "I say, Watson, what o'clock is it?"

"Nearly eleven."