9 Down Close
"A fair dowry. Not more than is usual in my family."

"And this, of course, remains to you, since the marriage is a fait accompli?"

"I really have made no inquiries on the subject."

"Very naturally not. Did you see Miss Doran on the day before the wedding?"

"Yes."

"Was she in good spirits?"