48 Ventnor Terrace
"I cannot say that I have."

"The lady could not move her bed. It must always be in the same relative position to the ventilator
and to the rope--or so we may call it, since it was clearly never meant for a bell-pull."

"Holmes," I cried, "I seem to see dimly what you are hinting at. We are only just in time to prevent
some subtle and horrible crime."