31 Balcombe Close
"No. They were found floating near the margin by a park-keeper. They have been identified as her
clothes, and it seemed to me that if the clothes were there the body would not be far off."

"By the same brilliant reasoning, every man's body is to be found in the neighbourhood of his
wardrobe. And pray what did you hope to arrive at through this?"

"At some evidence implicating Flora Millar in the disappearance."