24 Badgers Dene
"I wrote my articles and thought little more of the matter until, some time later, I backed a bill
for a friend and had a writ served upon me for £25. I was at my wit's end where to get the money,
but a sudden idea came to me. I begged a fortnight's grace from the creditor, asked for a holiday
from my employers, and spent the time in begging in the City under my disguise. In ten days I had
the money and had paid the debt.

"Well, you can imagine how hard it was to settle down to arduous work at £2 a week when I knew that
I could earn as much in a day by smearing my face with a little paint, laying my cap on the ground,
and sitting still. It was a long fight between my pride and the money, but the dollars won at last,
and I threw up reporting and sat day after day in the corner which I had first chosen, inspiring
pity by my ghastly face and filling my pockets with coppers. Only one man knew my secret. He was the
keeper of a low den in which I used to lodge in Swandam Lane, where I could every morning emerge as
a squalid beggar and in the evenings transform myself into a well-dressed man about town. This
fellow, a Lascar, was well paid by me for his rooms, so that I knew that my secret was safe in his
possession.